CAN ID 원래대로 함
엔코더 옵셋대로 동작되는거 확인, 각도 단위롤 flot에서 uint16_t로 바꿔 계산하게 함

+++ Inc/hx_WristSensor.h
@@ -163,7 +163,7 @@ typedef struct tag_WRIST{
 	struct {
 		uint16_t *pValue;
 		uint16_t value;
-		float offset;
+		int32_t offset;
 		uint16_t angle; //x100
 	}m_Enc[2];
 	struct {
